% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{boundary}
\alias{boundary}
\title{Boundaries}
\usage{
boundary(district, host, port, user, password, dbname, enctype)
}
\arguments{
\item{district}{City name, character}

\item{host}{Host of the server, character}

\item{port}{Port of the server, numeric}

\item{user}{User of the server, character}

\item{password}{Password of the server, character}

\item{dbname}{Database name of city, character}

\item{enctype}{The encoding type, gbk or utf8 or else ? character}
}
\value{
A list:

(1) administration boundary, data frame

(2) house boundary, data frame
}
\description{
Extract the administration bondary and effective position restricted by existed houses, of a city
}

