% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_city.R
\name{hp_city}
\alias{hp_city}
\title{House Price calculation of each city}
\usage{
hp_city(district, host, port, user, password, dbname, startmon, endmon, resol, outpath,
 sys)
}
\arguments{
\item{district}{City name, character}

\item{host}{Host of the server, character}

\item{port}{Port of the server, numeric}

\item{user}{User of the server, character}

\item{password}{Password of the server, character}

\item{dbname}{Database name of city, character}

\item{startmon}{The firt month in the form such as "200606", character}

\item{endmon}{The last month in the form as param startmon, character}

\item{resol}{Mesh resolution, unit: meter, numeric}

\item{sys}{The system type, Linux or Wins, defines the encoding type of configure file here, character}
}
\value{
Magnitude, Link and Year-over-year distibution of House Price; price level; minmax magnitude
}
\description{
Calculate every city results in spetial database, based on Odinary Krigging interpolation.
}
\details{
THe outputs mainly contains Altitude, Link and Year-over-year distibution of the price.
}

