% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessPost.R
\name{preprocessPost}
\alias{preprocessPost}
\title{Preprocess}
\usage{
preprocessPost(district, host, port, user, password, dbname, mon, enctype,
  outpath)
}
\arguments{
\item{district}{City name, character}

\item{host}{Host of the server, character}

\item{port}{Port of the server, numeric}

\item{user}{User of the server, character}

\item{password}{Password of the server, character}

\item{dbname}{Database name of city, character}

\item{mon}{Current month in the form as param startmon, character}

\item{enctype}{The encoding type, gbk or utf8 or else ? character}
}
\value{
A list:

(1) preproccessed data, data frame

(2) the first month conform to some regularity, character

(3) the last month conform to some regularity, character

Note: here, the total number of sample/month should be greater than 10.
}
\description{
Extract useful data frome database, and organize them to a needed format
}

